+++ src/flag_gems/ops/div.py
@@ -302,50 +302,3 @@ def div_mode_(A, B, rounding_mode=None):
     else:
         msg = f"div expected rounding_mode to be one of None, 'trunc', or 'floor' but found {rounding_mode}."
         raise ValueError(msg)
-
-
-@triton.jit
-def _remainder(x, y):
-    r = x % y
-    c1 = r != 0
-    c2 = (x < 0) ^ (y < 0)
-    return tl.where(c1 & c2, r + y, r)
-
-
-@pointwise_dynamic(promotion_methods=[(0, 1, "DEFAULT")])
-@triton.jit
-def rem_tt(x, y):
-    return _remainder(x, y)
-
-
-@pointwise_dynamic(is_tensor=[True, False], promotion_methods=[(0, 1, "DEFAULT")])
-@triton.jit
-def rem_ts(x, y):
-    return _remainder(x, y)
-
-
-@pointwise_dynamic(is_tensor=[False, True], promotion_methods=[(0, 1, "DEFAULT")])
-@triton.jit
-def rem_st(x, y):
-    return _remainder(x, y)
-
-
-def remainder(A, B):
-    logger.debug("GEMS REMAINDER")
-    if isinstance(A, torch.Tensor) and isinstance(B, torch.Tensor):
-        return rem_tt(A, B)
-    elif isinstance(A, torch.Tensor):
-        return rem_ts(A, B)
-    elif isinstance(B, torch.Tensor):
-        return rem_st(A, B)
-    else:
-        # Both scalar
-        return torch.tensor(A % B)
-
-
-def remainder_(A, B):
-    logger.debug("GEMS REMAINDER_")
-    if isinstance(B, torch.Tensor):
-        return rem_tt(A, B, out0=A)
-    else:
-        return rem_ts(A, B, out0=A)
